session10/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
bleach
passlib[bcrypt]
sqlalchemy
psycopg2-binary




